pyproject.toml
[project]
name = "azure-infra-cli"
version = "0.1.0"
description = "CLI tool for deploying Azure infrastructure using Bicep templates"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "azure-identity>=1.17",
    "azure-mgmt-resource>=23.0",
    "websockets>=16.0",
    "python-dotenv>=1.2.1",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5",
    "pytest>=8.0",
    "pytest-asyncio>=1.3.0",
    "mypy>=1.10",
    "pre-commit>=3.7",
]
whisper = ["faster-whisper>=1.0"]

[project.scripts]
azure-infra-cli = "infra_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/infra_cli", "src/gateway"]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
strict = true
python_version = "3.13"
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src/gateway"]
asyncio_mode = "auto"

